+++ test/helper.rb
@@ -13,7 +13,7 @@ require 'geoip'
 require 'coverage_loader'
 require 'vcr_loader'
 require 'test/unit'
-require 'mocha/setup'
+require 'mocha/test_unit'
 require 'net/http'
 
 require File.join(File.dirname(__FILE__), '../lib/geokit.rb')