+++ test/test_multi_ip_geocoder.rb
@@ -30,7 +30,9 @@ class MultiIpGeocoderTest < BaseGeocoderTest #:nodoc: all
   def test_invalid_provider
     temp = Geokit::Geocoders.ip_provider_order
     Geokit::Geocoders.ip_provider_order = [:bogus]
-    assert_equal @failure, Geokit::Geocoders::MultiGeocoder.geocode(@ip_address)
+    assert_raise Geokit::Geocoders::NoSuchGeocoderError do
+      Geokit::Geocoders::MultiGeocoder.geocode(@ip_address)
+    end
     Geokit::Geocoders.ip_provider_order = temp
   end
 end